[fix] header struct
